refactor: PR#15レビュー推奨改善を適用する
- useRef パターンで onCountsUpdate を保持し useEffect 依存配列から除去（アニメーションループの不要な再起動を防止）
- SpeciesCounts 型と createEmptySpeciesCounts() を lib/speciesUtils.ts に移動
- useState の遅延初期化に変更（TerminalWindow）
- SpriteIcon props を Discriminated Union に変更（species/isShark の排他を型で保証）
- drawSpriteIcon にガード節を追加（空スプライトへの配列アクセスを防止）
- import React を削除して named import のみに統一（PopulationPanel）
